0013030: Withdrawn SOP - no access message
when a SOP status is changed to 'withdrawn' and a standard user (one with no admin rights) clicks on the SOP, the message displayed lists all SOP IDs with withdrawn status (as per attached screenshot). We only need this to display the SOP ID for the one that has been clicked.
Also, message states 'you are not a reviewer of this SOP'. Please can the message state the following only:
"You do not have access to this SOP. This SOP is no longer active"