MORTAR also doesn’t record the date when a reminde
MORTAR also doesn’t record the date when a reminde Rebecca Brown
Mon 03/02/2020 16:19

In MORTAR, there used to be a report available called ‘Training Record Reminder List’ but it doesn’t seem to be there anymore. It was under the ‘Training Record’ drop down list.

MORTAR also doesn’t record the date when a reminder was sent anymore – only the created and completed dates.

Please could you let me know if you are aware of this or if I need to raise a Mantis?